backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
dj-database-url==2.1.0
Django==3.2.22
django-cors-headers==4.0.0
django-phonenumber-field==7.1.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-core==2.3.3
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
jsonschema-specifications==0.1.3
multidict==6.0.4
packaging==23.2
phonenumbers==8.13.23
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==4.24.4
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyrsistent==0.19.3
pyTelegramBotAPI==4.14.0
python-decouple==3.8
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.8.9
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
whitenoise==6.4.0
yarl==1.9.2
zipp==3.15.0
